// frontend/app/services/analytics/types.gen.ts
export type DailyMeanPercentageItem = {
    /**
     * Mean Percentage
     */
    mean_percentage: number;
    /**
     * Day
     */
    day: string;
};

/**
 * Frequency
 */
export type Frequency = 'hour' | 'day';

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * HourlyAggregationItem
 */
export type HourlyAggregationItem = {
    /**
     * Occupied Count
     */
    occupied_count: number;
    /**
     * Hour
     */
    hour: string;
};

/**
 * HourlyMeanPercentageItem
 */
export type HourlyMeanPercentageItem = {
    /**
     * Mean Percentage
     */
    mean_percentage: number;
    /**
     * Hour
     */
    hour: string;
};

/**
 * MeanPercentageDto
 */
export type MeanPercentageDto = {
    frequency: Frequency;
    period_range: PeriodRange;
    /**
     * Mean Percentages
     */
    mean_percentages: Array<HourlyMeanPercentageItem> | Array<DailyMeanPercentageItem>;
    /**
     * Highest Mean Datetime
     */
    highest_mean_datetime: string | null;
    /**
     * Lowest Mean Datetime
     */
    lowest_mean_datetime: string | null;
};

/**
 * PeriodRange
 */
export type PeriodRange = {
    /**
     * Start Date
     */
    start_date: string;
    /**
     * End Date
     */
    end_date: string;
};